Renamed folder get-tasks to get-tasks-homepage
